Drop sounds.ramfs and have asterisk find the appropriate directories.

+++ bin/osdial-astsnds-ramfs.sh
@@ -29,50 +29,27 @@ ASTDIR=/var/lib/asterisk
 
 RAMMNT=`mount | grep "${RAMDIR}"`
 
-if [ -d "${ASTDIR}/sounds.ramfs" -a -d "${ASTDIR}/sounds" -a ! -L "${ASTDIR}/sounds"]; then
+if [ -d "${ASTDIR}/sounds.ramfs" ]; then
 	# Update our sound files on the harddisk and remove sounds directory..
 	[ $DB -gt 0 ] && echo -e "\nMerge duplicate sounds directories." || :
-	if [ -L "${ASTDIR}/sounds/sounds"]; then
-		rm -f ${ASTDIR}/sounds/sounds
-	fi
+	[ -L "${ASTDIR}/sounds/sounds" ] && rm -f ${ASTDIR}/sounds/sounds || :
+	[ -L "${ASTDIR}/sounds" ] && rm -f ${ASTDIR}/sounds || :
 	if [ -d "${RAMDIR}/sounds" ]; then
-		cp -rf ${RAMDIR}/sounds/* ${ASTDIR}/sounds.ramfs
+		yes | cp -auf ${RAMDIR}/sounds/* ${ASTDIR}/sounds.ramfs
 		rm -rf ${RAMDIR}/sounds
 	fi
-	cp -rf ${ASTDIR}/sounds/* ${ASTDIR}/sounds.ramfs
+	yes | cp -auf ${ASTDIR}/sounds/* ${ASTDIR}/sounds.ramfs
 	rm -rf ${ASTDIR}/sounds
 	mv ${ASTDIR}/sounds.ramfs ${ASTDIR}/sounds
-	if [ -d "${ASTDIR}/OSDprompts" ]; then
-		cp ${ASTDIR}/OSDprompts/* ${ASTDIR}/sounds
-	fi
+	[ -d "${ASTDIR}/OSDprompts" ] && yes | cp -auf ${ASTDIR}/OSDprompts/* ${ASTDIR}/sounds || :
 fi
 
-# RAMDIR is NOT mounted...we better put everything back.
-if [ -z "$RAMMNT" ]; then
-	if [ -d "${ASTDIR}/sounds.ramfs" ]; then
-		# If sounds is a symlink, remove.
-		[ -L "${ASTDIR}/sounds" ] && rm -f ${ASTDIR}/sounds || :
-
-		# If sounds is a directory copy over it and remove sounds.ramfs
-		if [ -d "${ASTDIR}/sounds" ]; then
-			cp -rf ${ASTDIR}/sounds.ramfs/* ${ASTDIR}/sounds
-			rm -rf ${ASTDIR}/sounds.ramfs
-
-		# If sounds does not exist, move sounds.ramfs over it.
-		elif [ ! -e "${ASTDIR}/sounds" ]; then
-			mv ${ASTDIR}/sounds.ramfs ${ASTDIR}/sounds
-		fi
-		if [ -d "${ASTDIR}/OSDprompts" ]; then
-			cp ${ASTDIR}/OSDprompts/* ${ASTDIR}/sounds
-		fi
-	fi
-
 # RAMDIR is mounted...lets move the sounds to RAMFS.
-elif [ -n "$RAMMNT" ]; then
-	if [ ! -e "${ASTDIR}/sounds.ramfs" -a ! -e "${ASTDIR}/sounds" -a ! -e "${RAMDIR}/sounds" ]; then
+if [ -n "$RAMMNT" ]; then
+	if [ ! -e "${ASTDIR}/sounds" -a ! -e "${RAMDIR}/sounds" ]; then
 		[ $DB -gt 0 ] && echo -e "\nERROR: Fatal error, all sounds directories missing.!\n" || :
 		exit 1
-	elif [ ! -d "${ASTDIR}/sounds.ramfs" -a -d "${ASTDIR}/sounds" ]; then
+	elif [ -d "${ASTDIR}/sounds" -a ! -d "${RAMDIR}/sounds" ]; then
 		# Check for available space.
 		[ $DB -gt 0 ] && echo -e "\nChecking for available/required memory." || :
 		RAMAVAIL=`df -k | grep "${RAMDIR}" | awk '{ print $4 }'`
@@ -86,34 +63,27 @@ elif [ -n "$RAMMNT" ]; then
 		fi
 		# Initial creation.
 		[ $DB -gt 0 ] && echo -e "\nMoving sound files to RAMFS." || :
-		mv ${ASTDIR}/sounds ${ASTDIR}/sounds.ramfs
-		mkdir -p ${RAMDIR}/sounds
 		if [ -d "${ASTDIR}/OSDprompts" ]; then
-			cp ${ASTDIR}/OSDprompts/* ${ASTDIR}/sounds.ramdir
+			yes | cp -auf ${ASTDIR}/OSDprompts/* ${ASTDIR}/sounds
+			yes | cp -auf ${ASTDIR}/sounds/851* ${ASTDIR}/OSDprompts
 		fi
-		cp -rf ${ASTDIR}/sounds.ramfs/* ${RAMDIR}/sounds
-	elif [ -d "${ASTDIR}/sounds.ramfs" -a -d "${RAMDIR}/sounds" ]; then
+		mkdir -p ${RAMDIR}/sounds
+		yes | cp -auf ${ASTDIR}/sounds/* ${RAMDIR}/sounds
+	elif [ -d "${ASTDIR}/sounds" -a -d "${RAMDIR}/sounds" ]; then
 		# Update our sound files on the harddisk.
 		[ $DB -gt 0 ] && echo -e "\nUpdate sound files on HD." || :
 		if [ -d "${ASTDIR}/OSDprompts" ]; then
-			cp ${ASTDIR}/OSDprompts/* ${RAMDIR}/sounds
+			yes | cp -auf ${ASTDIR}/OSDprompts/* ${ASTDIR}/sounds
+			yes | cp -auf ${ASTDIR}/sounds/851* ${ASTDIR}/OSDprompts
 		fi
-		cp -rf ${RAMDIR}/sounds/* ${ASTDIR}/sounds.ramfs
-	elif [ -d "${ASTDIR}/sounds.ramfs" -a ! -d "${RAMDIR}/sounds" ]; then
-		# Recover RAMFS sound files.
-		[ $DB -gt 0 ] && echo -e "\nUpdate sound files on RAMFS." || :
-		mkdir -p ${RAMDIR}/sounds
-		if [ -d "${ASTDIR}/OSDprompts" ]; then
-			cp ${ASTDIR}/OSDprompts/* ${ASTDIR}/sounds.ramfs
-		fi
-		cp -rf ${ASTDIR}/sounds.ramfs/* ${RAMDIR}/sounds
+		yes | cp -auf ${ASTDIR}/sounds/* ${RAMDIR}/sounds
+		yes | cp -auf ${RAMDIR}/sounds/* ${ASTDIR}/sounds
 	fi
 fi
 
 # Clean-up
-[ ! -L "${ASTDIR}/sounds" -a -d "${ASTDIR}/sounds.ramfs" -a -d "${RAMDIR}/sounds" ] && ln -s ${RAMDIR}/sounds ${ASTDIR}/sounds || :
-[ -d "${ASTDIR}/sounds.ramfs" ] && chown -R asterisk:asterisk ${ASTDIR}/sounds.ramfs || :
-[ -d "${RAMDIR}/sounds" ]       && chown -R asterisk:asterisk ${RAMDIR}/sounds || :
-[ -d "${ASTDIR}/sounds" ]       && chown -R asterisk:asterisk ${ASTDIR}/sounds || :
+[ -d "${ASTDIR}/sounds" ] &&     chown -R asterisk:asterisk ${ASTDIR}/sounds || :
+[ -d "${ASTDIR}/OSDprompts" ] && chown -R asterisk:asterisk ${ASTDIR}/OSDprompts || :
+[ -d "${RAMDIR}/sounds" ] &&     chown -R asterisk:asterisk ${RAMDIR}/sounds || :
 
 [ $DB -gt 0 ] && echo -e "Done.\n" || :
